refactor: simplify entity and value object filenames
