+++ src/backend/src/routes/application/use-cases/list-routes.ts
@@ -1,5 +1,5 @@
 import { RouteRepository } from '../../domain/repositories/route-repository';
-import { Route } from '../../domain/entities/route-entity';
+import { Route } from '../../domain/entities/route';
 
 export class ListRoutesUseCase {
   constructor(private repository: RouteRepository) {}
