+++ src/backend/src/routes/application/use-cases/list-routes.usecase.test.ts
@@ -1,7 +1,7 @@
 import { ListRoutesUseCase } from './list-routes';
 import { RouteRepository } from '../../domain/repositories/route-repository';
-import { Route } from '../../domain/entities/route-entity';
-import { UUID } from '../../../shared/domain/value-objects/uuid-value-object';
+import { Route } from '../../domain/entities/route';
+import { UUID } from '../../../shared/domain/value-objects/uuid';
 
 describe('ListRoutesUseCase', () => {
   it('returns all routes from repository', async () => {
