+++ src/backend/src/routes/domain/events/route-generated.test.ts
@@ -1,10 +1,10 @@
 import { RouteGeneratedEvent } from "./route-generated";
-import { Route } from "../entities/route-entity";
-import { UUID } from "../../../shared/domain/value-objects/uuid-value-object";
-import { DistanceKm } from "../value-objects/distance-value-object";
-import { Duration } from "../value-objects/duration-value-object";
-import { Path } from "../value-objects/path-value-object";
-import { LatLng } from "../value-objects/lat-lng-value-object";
+import { Route } from "../entities/route";
+import { UUID } from "../../../shared/domain/value-objects/uuid";
+import { DistanceKm } from "../value-objects/distance";
+import { Duration } from "../value-objects/duration";
+import { Path } from "../value-objects/path";
+import { LatLng } from "../value-objects/lat-lng";
 
 describe("RouteGeneratedEvent", () => {
   it("should hold the generated route", () => {
