+++ src/backend/src/routes/domain/services/route-generator.ts
@@ -1,9 +1,9 @@
-import { Route } from "../entities/route-entity";
+import { Route } from "../entities/route";
 import { RouteRepository } from "../repositories/route-repository";
-import { DistanceKm } from "../value-objects/distance-value-object";
-import { Duration } from "../value-objects/duration-value-object";
-import { Path } from "../value-objects/path-value-object";
-import { UUID } from "../../../shared/domain/value-objects/uuid-value-object";
+import { DistanceKm } from "../value-objects/distance";
+import { Duration } from "../value-objects/duration";
+import { Path } from "../value-objects/path";
+import { UUID } from "../../../shared/domain/value-objects/uuid";
 import { fetchJson } from "../../interfaces/shared/utils";
 import { request as httpsRequest, RequestOptions } from "node:https";
