+++ src/backend/src/routes/domain/value-objects/lat-lng.ts
