+++ src/backend/src/routes/interfaces/sqs/worker-routes.ts
@@ -3,8 +3,8 @@ import { DynamoDBClient } from "@aws-sdk/client-dynamodb";
 import { SQSClient, SendMessageCommand } from "@aws-sdk/client-sqs";
 import { request as httpsRequest, RequestOptions } from "node:https";
 import { createHash } from "node:crypto";
-import { Path } from "../../domain/value-objects/path-value-object";
-import type { Route } from "../../domain/entities/route-entity";
+import { Path } from "../../domain/value-objects/path";
+import type { Route } from "../../domain/entities/route";
 import { DynamoRouteRepository } from "../../infrastructure/dynamodb/dynamo-route-repository";
 import { publishRoutesGenerated } from "../appsync-client";
 import { fetchJson, getGoogleKey } from "../shared/utils";
