+++ src/backend/src/shared/domain/value-objects/email.ts
