+++ src/backend/src/shared/domain/value-objects/uuid.ts
