+++ src/backend/src/users/domain/value-objects/age.test.ts
@@ -1,4 +1,4 @@
-import { Age } from './age-value-object';
+import { Age } from './age';
 
 describe('Age', () => {
   it('creates from valid number', () => {
